Extract timer state into useTimers hook
- Move all timer state, handlers, and effects from App.tsx into src/hooks/useTimers.ts
- App.tsx now only handles rendering
- No behavior change; all 108 tests still pass

Closes #13
Closes #16
